Add CS fixer
