+++ .travis.yml
@@ -23,3 +23,4 @@ install:
 
 script:
   - composer test
+  - composer cs
