+++ composer.json
@@ -23,7 +23,8 @@
         "container-interop/container-interop": "^1.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.2"
+        "phpunit/phpunit": "~5.2",
+        "squizlabs/php_codesniffer": "^2.3"
     },
     "autoload": {
         "psr-4": {
@@ -32,10 +33,16 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "ZfrShopifyTest\\": "tests/"
+            "ZfrShopifyTest\\": "test/"
         }
     },
     "scripts": {
+        "check": [
+            "@cs",
+            "@test"
+        ],
+        "cs": "phpcs",
+        "cs-fix": "phpcbf",
         "test": "phpunit"
     }
 }
