+++ src/Exception/ExceptionInterface.php
