+++ test/Model/ShopDomainTest.php
